+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `Animation::skinning_transforms` and `Animation::model_space_transforms` to xc3_model.
 * Added field `texcoord_name` to `ChannelAssignment::Texture` for xc3_model.
 * Added basic support for loading and rendering `.camdo` models from Xenoblade X.
+* Added support for exporting `.camdo` models with xc3_gltf.
 * Added support for Xenoblade X `.catex`, `.calut`, `.caavp`, and `.bmn` textures to xc3_tex.
 * Added support for uncompressed and zstd compressed `Xbc1` files.
 * Added rendering support for bone XYZ axes to xc3_wgpu.